# nodejs-2025/home-library-service/doc/api.yaml
openapi: 3.0.0
paths:
...
  /user:
    get:
      operationId: UserController_findAll
      parameters: []
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
      security:
        - bearer: []
      summary: Get all users
      tags:
        - User
    post:
      operationId: UserController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateDto'
      responses:
        '201':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      security:
        - bearer: []
      summary: Get user by id
      tags:
        - User
  /user/{id}:
    get:
      operationId: UserController_findByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      security:
        - bearer: []
      summary: Get user by id
      tags:
        - User
    put:
      operationId: UserController_updateByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdatePasswordDto'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      security:
        - bearer: []
      summary: Update user by id
      tags:
        - User
    delete:
      operationId: UserController_deleteByIdOrException
      parameters:
        - name: id
          required: true
          in: path
          description: UUID field
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Success request
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseDto'
      security:
        - bearer: []
      summary: Delete user by id
      tags:
        - User
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    HttpStatus:
      type: number
      enum:
        - 100
        - 101
        - 102
        - 103
        - 200
        - 201
        - 202
        - 203
        - 204
        - 205
        - 206
        - 207
        - 208
        - 210
        - 300
        - 301
        - 302
        - 303
        - 304
        - 307
        - 308
        - 400
        - 401
        - 402
        - 403
        - 404
        - 405
        - 406
        - 407
        - 408
        - 409
        - 410
        - 411
        - 412
        - 413
        - 414
        - 415
        - 416
        - 417
        - 418
        - 421
        - 422
        - 423
        - 424
        - 428
        - 429
        - 456
        - 500
        - 501
        - 502
        - 503
        - 504
        - 505
        - 507
        - 508
      description: Exception status code
    ExceptionResponseDto:
      type: object
      properties:
        message:
          type: string
          description: Exception message
          example: User not found
        path:
          type: string
          description: Endpoint path
          example: /user/123
        timestamp:
          format: date-time
          type: string
          description: Exception timestamp
        statusCode:
          description: Exception status code
          example: 404
          allOf:
            - $ref: '#/components/schemas/HttpStatus'
      required:
        - message
        - path
        - timestamp
        - statusCode
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          description: UUID field
          format: uuid
        createdAt:
          type: number
          description: User create date timestamp
        login:
          type: string
          description: User login
        updatedAt:
          type: number
          description: User update date timestamp
        version:
          type: number
          description: User entity version
          format: int32
      required:
        - id
        - createdAt
        - login
        - updatedAt
        - version
...
    UserCreateDto:
      type: object
      properties:
        login:
          type: string
          description: User login
        password:
          type: string
          description: User password
      required:
        - login
        - password
    UserUpdatePasswordDto:
      type: object
      properties:
        newPassword:
          type: string
          description: New user password
        oldPassword:
          type: string
          description: Old user password
      required:
        - newPassword
        - oldPassword