# nodejs-2025/home-library-service/src/modules/user/dto/user.create.dto.ts
import { ApiProperty, PickType } from '@nestjs/swagger';
import { IsNotEmpty, IsString } from 'class-validator';
import { IUserCreate } from '../../../models/user.model';
import { UserDto } from '../../../dto/models/user.dto';

export class UserCreateDto
  extends PickType(UserDto, ['login'] as const)
  implements IUserCreate
{
  @ApiProperty({ description: 'User password' })
  @IsNotEmpty()
  @IsString()
  password: string;
}